L5/L6/L7: compileJava is green
Remaining 1.21 API work:
- GeckoLib 4.9.2 preRender/reRender: RGBA floats -> packed int colour
- Model#renderToBuffer likewise; ModelPart.render follows
- MobEffect lifecycle: applyEffectTick returns boolean,
  isDurationEffectTick -> shouldApplyEffectTickThisTick,
  addAttributeModifiers(entity,..) -> onEffectStarted. 1.21 has no
  entity-aware removal hook and NeoForge dropped curative items, so the
  expiry callbacks moved to a MobEffectEvent.Expired/Remove listener
- SpawnPlacements.register went private -> RegisterSpawnPlacementsEvent,
  so the 15 entity init() methods take the event instead of running
  during common setup
- AbstractArrow carries a pickup ItemStack and declares
  getDefaultPickupItem; canChangeDimensions takes both levels;
  getMyRidingOffset and MeleeAttackGoal#getAttackReachSqr are gone
- finalizeSpawn and EntityType#create lost their CompoundTag parameter
- DimensionSpecialEffects: renderClouds gained a modelViewMatrix and
  renderSky's PoseStack became a Matrix4f (rebuilt into a PoseStack so
  the registered sky predicates keep their argument contract)
- MenuScreens.register went private -> RegisterMenuScreensEvent

Data: 1.21 singularised loot_tables/entity_types and NeoForge replaced
the forge namespace. ImaginaryPurple's canBreatheUnderwater override is
now the #minecraft:can_breathe_under_water tag, since the method is final.

GUI: the radial menu is a plain Screen, which on 1.21 would inherit the
new blurred background; it now overrides renderBackground to a no-op.
ArmAnimationEditorScreen drops its redundant manual renderBackground call
(AbstractContainerScreen's own override is already the unblurred one).

Fixes 3 sites where the earlier setKnockback removal left a dangling
receiver expression behind.

+++ src/main/java/net/efkrdnz/jjkstrongest/procedures/ChargeRingShapeProcedure.java
@@ -26,9 +26,9 @@ public class ChargeRingShapeProcedure {
 		if (bufferBuilder == null)
 			return;
 		if (format == DefaultVertexFormat.POSITION_COLOR) {
-			bufferBuilder.addVertex(x, y, z).setColor(color);
+			bufferBuilder.addVertex((float) x, (float) y, (float) z).setColor(color);
 		} else if (format == DefaultVertexFormat.POSITION_TEX_COLOR) {
-			bufferBuilder.addVertex(x, y, z).setUv(u, v).setColor(color);
+			bufferBuilder.addVertex((float) x, (float) y, (float) z).setUv(u, v).setColor(color);
 		}
 	}
